First updates into GIT

+++ src/bioLockJ/metagenome/RunMultipleKraken_MultiplePerCore.java
@@ -36,22 +36,16 @@ public class RunMultipleKraken_MultiplePerCore extends BioLockJExecutor
 	{	
 		BioLockJUtils.requireString(cReader, ConfigReader.CLUSTER_BATCH_COMMAND);
 		BioLockJUtils.requireExistingDirectory(cReader, ConfigReader.PATH_TO_INPUT_RDP_FASTA_DIRECTORY);
-		BioLockJUtils.requireExistingDirectory(cReader, ConfigReader.PATH_TO_KRAKEN_OUTPUT_DIRECTORY);
 		BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_BINARY);
 		BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_DATABASE);
-		BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_SCRIPT_DIR);
 		BioLockJUtils.requirePositiveInteger(cReader, ConfigReader.NUMBER_OF_JOBS_PER_CORE);
 	}
 	
-	private File makeNewRunFile( File rdpScriptDir, BufferedWriter allWriter,
-					String clusterCommand, int countNum) throws Exception
+	private File makeNewRunFile(String scriptDir, BufferedWriter allWriter, 
+			String clusterCommand, int countNum) throws Exception
 	{
-		File runFile = new File(rdpScriptDir.getAbsoluteFile() + File.separator + "run_" + 
-				countNum + "_" + System.currentTimeMillis() +  ".sh");
-	
+		File runFile = new File(scriptDir + "run_" + countNum + "_" + System.currentTimeMillis() +  ".sh");
 		this.scriptFiles.add(runFile);
-		
-		
 		allWriter.write(clusterCommand + " " +  runFile.getAbsolutePath() + "\n"  );
 		allWriter.flush();
 	
@@ -62,14 +56,9 @@ public class RunMultipleKraken_MultiplePerCore extends BioLockJExecutor
 		throws Exception
 	{
 		File touchFile = new File(runFile.getAbsolutePath() + FINISHED_SUFFIX );
-		
-		if( touchFile.exists())
-			touchFile.delete();
-		
+		if( touchFile.exists()) touchFile.delete();
 		aWriter.write("touch " + touchFile.getAbsolutePath() + "\n");
-		
 		aWriter.flush();  aWriter.close();
-		
 	}
 	
 	@Override
@@ -77,25 +66,22 @@ public class RunMultipleKraken_MultiplePerCore extends BioLockJExecutor
 	{
 		String clusterCommand = BioLockJUtils.requireString(cReader, ConfigReader.CLUSTER_BATCH_COMMAND);
 	 	File fastaInDir =  BioLockJUtils.requireExistingDirectory(cReader, ConfigReader.PATH_TO_INPUT_RDP_FASTA_DIRECTORY);
-		File krakenOutDir =  BioLockJUtils.requireExistingDirectory(cReader, ConfigReader.PATH_TO_KRAKEN_OUTPUT_DIRECTORY);
 		File krakenBinary =  BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_BINARY);
-		File krakenScriptDir =  BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_SCRIPT_DIR);
-		
 		File krakenDatabase = BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_DATABASE);
 		int numJobsPerCore = BioLockJUtils.requirePositiveInteger(cReader, ConfigReader.NUMBER_OF_JOBS_PER_CORE);
 		String clusterParams = BioLockJUtils.getStringOrNull(cReader, ConfigReader.CLUSTER_PARAMS);
 		
+		//File krakenScriptDir =  BioLockJUtils.requireExistingFile(cReader, ConfigReader.PATH_TO_KRAKEN_SCRIPT_DIR);
+		//File krakenOutDir =  BioLockJUtils.requireExistingDirectory(cReader, ConfigReader.PATH_TO_KRAKEN_OUTPUT_DIRECTORY);
+		
 		String[] files = fastaInDir.list();
-	
-		this.runAllFile = 
-				new File(
-						krakenScriptDir.getAbsoluteFile() + File.separator + "runAll.sh");
+
+		this.runAllFile = new File(getScriptDir(cReader) + File.separator + "runAll.sh");
 		
 		BufferedWriter allWriter = new BufferedWriter(new FileWriter(runAllFile));
-		
 		int countNum=0;
 		int numToDo = numJobsPerCore;
-		File runFile = makeNewRunFile(krakenScriptDir, allWriter, clusterCommand, countNum);
+		File runFile = makeNewRunFile(getScriptDir(cReader), allWriter, clusterCommand, countNum);
 		BufferedWriter aWriter = new BufferedWriter(new FileWriter(runFile));
 		
 		for(String s : files)
@@ -103,37 +89,33 @@ public class RunMultipleKraken_MultiplePerCore extends BioLockJExecutor
 			countNum++;
 			File fastaFile = new File(fastaInDir.getAbsolutePath() + File.separator + s);
 			
-			File krakenOutFile = new File(krakenOutDir.getAbsolutePath() + File.separator + s  + "toKraken.txt");
-			File krakenTranslateFile = 
-					new File(krakenOutDir.getAbsolutePath() + File.separator + s  + "toKrakenTranslate.txt");
+			//File krakenOutFile = new File(krakenOutDir.getAbsolutePath() + File.separator + s  + "toKraken.txt");
+			//File krakenTranslateFile = new File(krakenOutDir.getAbsolutePath() + File.separator + s  + "toKrakenTranslate.txt");
+			
+			String krakenOutput = getOutputDir(cReader) + s + "toKraken.txt";
+			String krakenTranslate = getOutputDir(cReader) + s + "toKrakenTranslate.txt";
 			
 			if( clusterParams != null)
 				aWriter.write(clusterParams + "\n");
 			
-			aWriter.write(krakenBinary.getAbsolutePath() + " -db " +  
-					krakenDatabase.getAbsolutePath()  + " --output " + krakenOutFile.getAbsolutePath() +
-					" " +  fastaFile+ "\n" );
+			aWriter.write(krakenBinary.getAbsolutePath() + " --db " +  krakenDatabase.getAbsolutePath()  + 
+					" --output " + krakenOutput + " " +  fastaFile + "\n" );
+			aWriter.write("if [ $? –eq 0 ]; then \n" );
+			aWriter.write("    " + krakenBinary.getAbsolutePath() + "-translate " + " --db " +  
+					krakenDatabase.getAbsolutePath()  + " " + krakenOutput + " > " + krakenTranslate + "\n" );
+			aWriter.write("else touch " + getOutputDir(cReader) + s + FAILED_TO_PROCESS + " fi \n" );
 			
-			
-			aWriter.write(krakenBinary.getAbsolutePath() + "-translate " + " -db " +  
-					krakenDatabase.getAbsolutePath()  + " " + krakenOutFile.getAbsolutePath() +
-					" > " + krakenTranslateFile + "\n" );
-			
-			
-			numToDo--;
-			
-			if( numToDo == 0 )
+			//numToDo--; MS move "--" into in conditional before param to subtract before check			
+			if( --numToDo == 0 )
 			{
 				numToDo = numJobsPerCore;
 				closeARunFile(aWriter, runFile);
-				runFile = makeNewRunFile(krakenScriptDir, allWriter, clusterCommand,countNum);
+				runFile = makeNewRunFile(getScriptDir(cReader), allWriter, clusterCommand, countNum);
 				aWriter = new BufferedWriter(new FileWriter(runFile));
 			}
-			
 		}
 
 		closeARunFile(aWriter, runFile);
 		allWriter.flush();  allWriter.close();
-
 	}
 }
